refactor(core): move GET users/:userId/logs to routes/log.ts
