+++ packages/core/src/routes/admin-user.test.ts
@@ -3,7 +3,6 @@ import pick from 'lodash.pick';
 
 import { mockUser, mockUserList, mockUserListResponse, mockUserResponse } from '@/__mocks__';
 import { encryptUserPassword } from '@/lib/user';
-import { UserLogCondition } from '@/queries/log';
 import { findRolesByRoleNames } from '@/queries/roles';
 import {
   hasUser,
@@ -57,23 +56,6 @@ jest.mock('@/lib/user', () => ({
     passwordEncrypted: 'password',
     passwordEncryptionMethod: 'Argon2i',
   })),
-}));
-
-const mockUserLogs = [{ id: 1 }, { id: 2 }];
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const countUserLogs = jest.fn(async (condition: UserLogCondition) => ({
-  count: mockUserLogs.length,
-}));
-const findUserLogs = jest.fn(
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  async (limit: number, offset: number, condition: UserLogCondition) => mockUserLogs
-);
-
-jest.mock('@/queries/log', () => ({
-  countUserLogs: async (condition: UserLogCondition) => countUserLogs(condition),
-  findUserLogs: async (limit: number, offset: number, condition: UserLogCondition) =>
-    findUserLogs(limit, offset, condition),
 }));
 
 jest.mock('@/queries/roles', () => ({
@@ -407,27 +389,5 @@ describe('adminUserRoutes', () => {
     });
     await userRequest.delete(`/users/${arbitraryUserId}/identities/${arbitraryConnectorId}`);
     expect(deleteUserIdentity).toHaveBeenCalledWith(arbitraryUserId, arbitraryConnectorId);
-  });
-
-  it('GET /users/:id/logs should call countUserLogs and findUserLogs with correct parameters', async () => {
-    const userId = 'userIdValue';
-    const applicationId = 'foo';
-    const logType = 'SignInUsernamePassword';
-    const page = 1;
-    const pageSize = 5;
-
-    await userRequest.get(
-      `/users/${userId}/logs?applicationId=${applicationId}&logType=${logType}&page=${page}&page_size=${pageSize}`
-    );
-    expect(countUserLogs).toHaveBeenCalledWith({ userId, applicationId, logType });
-    expect(findUserLogs).toHaveBeenCalledWith(5, 0, { userId, applicationId, logType });
-  });
-
-  it('GET /users/:id/logs should return correct response', async () => {
-    const userId = 'userIdValue';
-    const response = await userRequest.get(`/users/${userId}/logs`);
-    expect(response.status).toEqual(200);
-    expect(response.body).toEqual(mockUserLogs);
-    expect(response.header).toHaveProperty('total-number', `${mockUserLogs.length}`);
   });
 });
